requirements.txt
requests
PyPDF2
ebooklib
beautifulsoup4
python-dotenv
tqdm
pydub
